change user migration

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -35,6 +35,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('users');
+        }
     }
 };
